refactor(Shared.Configuration): restructure flat Configuration types into Shared.Configuration.<Sub>; split into per-type files
### Type relocation (closes second half of #17)

The four sibling Shared.* configuration types are now nested under
Shared.Configuration as a coherent tree:

  Shared.CrawlerConfiguration         -> Shared.Configuration.Crawler
  Shared.ChangeDetectionConfiguration -> Shared.Configuration.ChangeDetection
  Shared.OutputConfiguration          -> Shared.Configuration.Output
  Shared.OutputConfiguration.OutputFormat -> Shared.Configuration.Output.Format
  Shared.DiscoveryMode                -> Shared.Configuration.DiscoveryMode

Shared.Configuration (the aggregate struct) keeps its name and serves as
both the type AND the namespace — its three properties (crawler /
changeDetection / output) now refer to the nested types.

### One type per file (task #21)

Was one 247-line file with 5 mixed types; now six single-type files:

- Shared.Configuration.swift           — aggregate struct
- Shared.Configuration.Crawler.swift
- Shared.Configuration.ChangeDetection.swift
- Shared.Configuration.Output.swift
- Shared.Configuration.Output.Format.swift  (nested)
- Shared.Configuration.DiscoveryMode.swift

### External caller sweep (24+ files)

Across Sources/ and Tests/:

- Shared.CrawlerConfiguration         -> Shared.Configuration.Crawler
- Shared.ChangeDetectionConfiguration -> Shared.Configuration.ChangeDetection
- Shared.OutputConfiguration          -> Shared.Configuration.Output
- Shared.DiscoveryMode                -> Shared.Configuration.DiscoveryMode

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass

Closes the Shared.Configuration half of task #17. Combined with #418
(Shared.Core.ToolError), #17 is now complete.
