+++ Packages/Sources/Shared/Configuration/Shared.Configuration.Output.swift
@@ -1,0 +1,20 @@
+import Foundation
+import SharedConstants
+
+// MARK: - Shared.Configuration.Output
+
+extension Shared.Configuration {
+    /// Configuration for output format
+    public struct Output: Codable, Sendable {
+        public let format: Format
+        public let includeMarkdown: Bool
+
+        public init(
+            format: Format = .json,
+            includeMarkdown: Bool = false
+        ) {
+            self.format = format
+            self.includeMarkdown = includeMarkdown
+        }
+    }
+}
